+++ src/light/UniformPointLightSource.java
@@ -9,17 +9,17 @@ import java.util.*;
 public class UniformPointLightSource extends LightSource{
     private Random rndr = new Random();
     
-    public UniformPointLightSource(double[] position, double[] rotation,double[] color){
+    public UniformPointLightSource(float[] position, float[] rotation,float[] color){
         super(position, rotation, color);
     }
     
-    public double[] getNextBeam(){
-        double[] r = new double[5];
+    public float[] getNextBeam(){
+        float[] r = new float[5];
         r[0] = position[0];
         r[1] = position[1];
         r[2] = position[2];
-        r[3] = rndr.nextDouble() * 360.0;
-        r[4] = rndr.nextDouble() * 360.0;
+        r[3] = rndr.nextFloat() * 360;
+        r[4] = rndr.nextFloat() * 360;
         return r;
     }
 }
